+++ src/tewi/service/search/tpb_provider.py
@@ -129,6 +129,7 @@ class TPBProvider(BaseSearchProvider):
                 provider=self.short_name,
                 provider_id=self.id(),
                 page_url=page_url,
+                torrent_link=None,
                 fields=fields
             )