+++ src/tewi/service/search/yts_provider.py
@@ -222,6 +222,7 @@ class YTSProvider(BaseSearchProvider):
                 provider=self.short_name,
                 provider_id=self.id(),
                 page_url=page_url,
+                torrent_link=None,
                 fields=fields
             )